💄design/디자인 변경에 따른 디자인 작업 및 파일 위치 변경
