--- src/components/DetailStationLogo.tsx
@@ -1,5 +1,0 @@
-function DetailStationLogo() {
-  return <div className="detail-station-logo">MOOVE</div>;
-}
-
-export default DetailStationLogo;
