+++ src/detail/DetailStationResult.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import axios from 'axios';
 // components
-import '../styles/DetailStationResultStyle.scss';
+import '../styles/detail_result_style.scss';
 import { MdArrowBackIos, MdArrowForwardIos } from 'react-icons/md';
 import { currentStationAtom, stationInfoAtom } from '../recoil/atoms';
 import { ISubwayProps, ISubwayArriveInfo } from '../interfaces/interfaces';
@@ -15,7 +15,6 @@ export default function DetailStationResult() {
   const line5 = ['미사', '하남풍산', '하남시청', '하남검단산', '개롱', '거여'];
   // api address
   const arrivalAddress = 'http://swopenAPI.seoul.go.kr/api/subway/6267466b6b616e7437394670434955/json/realtimeStationArrival/0/5/';
-  // const tourAddress = 'http://openapi.seoul.go.kr:8088/4157654b4a616e743839737359467a/json/subwayTourInfo/1/5/';
 
   if (stationInfo) {
     if (Number(stationInfo[0]?.trnsitCo) > 1) {
@@ -50,10 +49,6 @@ export default function DetailStationResult() {
       .get<ISubwayProps>(`${arrivalAddress}${currentStation}`)
       .then((res) => setStationInfo(res.data.realtimeArrivalList))
       .catch((err) => console.log(err));
-    // axios
-    //   .get(`${stationIdAddress}`)
-    //   .then((res) => console.log(res.data.subwayTourInfo.row))
-    //   .catch((err) => console.log(err));
   }, [currentStation, setStationInfo]);
 
   return (
@@ -85,8 +80,8 @@ export default function DetailStationResult() {
         ))}
 
         <div className="exit-information">
-          <span className="exit-title">5번 출구</span>
-          <ul className="exit-list">
+          {/* <span className="exit-title">5번 출구</span> */}
+          {/* <ul className="exit-list">
             <li>
               <span>홍익대학교 서울 캠퍼스</span>
               <span>10분</span>
@@ -95,7 +90,7 @@ export default function DetailStationResult() {
               <span>홍익대학교 서울 캠퍼스</span>
               <span>10분</span>
             </li>
-          </ul>
+          </ul> */}
         </div>
       </div>
     </div>
