+++ src/header/DetailStationSearchWrap.tsx
@@ -1,6 +1,6 @@
 import DetailStationLineFilter from './DetailStationLineFilter';
 import DetailStationSearch from './DetailStationSearch';
-import '../styles/DetailStationSearch.scss';
+import '../styles/search_style.scss';
 
 function DetailStationSearchWrap() {
   return (
