+++ src/header/Logo.tsx
@@ -1,0 +1,5 @@
+function Logo() {
+  return <div className="detail-station-logo">MOOVE</div>;
+}
+
+export default Logo;
