+++ src/map/LineRect.tsx
