--- src/styles/DetailStationStyle.scss
@@ -1,21 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Squada+One&display=swap');
-
-.detail-station-wrap {
-  background-color: #f2f2f2;
-  border: 3px solid #d9d9d9;
-  border-top-right-radius: 5rem;
-  border-bottom-right-radius: 5rem;
-  width: 20%;
-  min-width: 450px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  .detail-station-contatiner {
-    .detail-station-logo {
-      width: 23rem;
-      font-family: 'Squada One', cursive;
-      font-size: 4rem;
-      letter-spacing: 1px;
-    }
-  }
-}
