+++ src/styles/breakpoints.scss
@@ -1,0 +1,4 @@
+// Breakpoints
+$breakpoint-mobile: 335px;
+$breakpoint-tablet: 758px;
+$breakpoint-desktop: 1024px;
