+++ src/styles/detail.scss
@@ -1,0 +1,39 @@
+@import '../styles/breakpoints.scss';
+
+.detail-station-wrap {
+  width: 100%;
+  height: 80%;
+  margin-top: 3rem;
+  @media screen and (max-width: #{$breakpoint-desktop}) {
+    margin-top: 0;
+  }
+  .detail-station-contatiner {
+    width: 100%;
+    height: 80vh;
+    display: flex;
+    @media screen and (max-width: #{$breakpoint-desktop}) {
+      height: 88vh;
+      flex-direction: column-reverse;
+      justify-content: space-between;
+      align-items: center;
+    }
+    .subway-map-border {
+      overflow: hidden;
+      width: 70%;
+      min-width: 1000px;
+      border: 3px solid var(--color-border);
+      border-radius: 3rem;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      @media screen and (max-width: #{$breakpoint-desktop}) {
+        min-width: 335px;
+        width: 90vw;
+        height: 60%;
+      }
+      .subway-map {
+        padding: 3rem;
+      }
+    }
+  }
+}
