+++ src/styles/detail_result_style.scss
@@ -1,10 +1,29 @@
+@import '../styles/breakpoints.scss';
+
 .station-result-wrap {
-  width: 100%;
-  overflow-y: scroll;
-  -ms-overflow-style: none; /* 인터넷 익스플로러 */
-  scrollbar-width: none; /* 파이어폭스 */
-  &::-webkit-scrollbar {
-    display: none; /* 크롬, 사파리, 오페라, 엣지 */
+  width: 25%;
+  min-width: 470px;
+  background-color: var(--color-sub);
+  border: 3px solid var(--color-border);
+  border-top-right-radius: 3rem;
+  border-bottom-right-radius: 3rem;
+  margin-right: 3rem;
+  // 스크롤
+  // overflow-y: scroll;
+  // -ms-overflow-style: none; /* 인터넷 익스플로러 */
+  // scrollbar-width: none; /* 파이어폭스 */
+  // &::-webkit-scrollbar {
+  //   display: none; /* 크롬, 사파리, 오페라, 엣지 */
+  // }
+  @media screen and (max-width: #{$breakpoint-desktop}) {
+    min-width: 335px;
+    width: 90vw;
+    height: 35%;
+    margin-right: 0;
+    border-top-left-radius: 3rem;
+    border-top-right-radius: 3rem;
+    border-bottom-right-radius: 0;
+    margin-top: 2rem;
   }
   .station-result-container {
     margin-top: 3rem;
@@ -76,6 +95,9 @@
         .next-station {
           display: flex;
           align-items: center;
+        }
+        @media screen and (max-width: #{$breakpoint-desktop}) {
+          width: 80%;
         }
       }
     }
