+++ src/styles/header.scss
@@ -1,0 +1,31 @@
+@import url('https://fonts.googleapis.com/css2?family=Squada+One&display=swap');
+@import '../styles/breakpoints.scss';
+
+.header-wrap {
+  width: 100%;
+  height: 20%;
+  @media screen and (max-width: #{$breakpoint-desktop}) {
+    height: 12vh;
+  }
+  .header-container {
+    width: 95%;
+    height: 5rem;
+    display: flex;
+    align-items: center;
+    padding: 1rem;
+    .logo {
+      min-width: max-content;
+      width: 25%;
+      margin-left: 1rem;
+      margin-right: 2rem;
+      background-color: #fff;
+      font-family: 'Squada One', cursive;
+      font-size: 4rem;
+      letter-spacing: 1px;
+      margin-right: 3rem;
+      @media screen and (max-width: #{$breakpoint-desktop}) {
+        display: none;
+      }
+    }
+  }
+}
