+++ src/styles/search_style.scss
@@ -1,20 +1,30 @@
+@import '../styles/breakpoints.scss';
+
 .DetailStationSearchWrap {
-  width: 23rem;
+  min-width: 1000px;
+  width: 70%;
+  height: max-content;
   display: flex;
-  flex-direction: column;
   align-items: center;
-  padding: 10% 0;
-  border-bottom: 3px solid rgba(170, 170, 170, 0.5);
+  @media screen and (max-width: #{$breakpoint-desktop}) {
+    min-width: 335px;
+    width: 100%;
+    justify-content: center;
+    padding: 0 1rem;
+  }
   .DetailStationSearch {
-    width: 23rem;
+    width: 45vw;
     height: 2rem;
     padding: 1%;
     display: flex;
     justify-content: space-between;
     align-items: center;
     background-color: white;
+    border: 3px solid var(--color-border);
     border-radius: 2rem;
+    margin-right: 3rem;
     position: relative;
+    z-index: 999;
     .DetailStationSearchInput {
       width: 100%;
       height: 1rem;
@@ -29,26 +39,26 @@
 
   .DetailStationSearchDropdown {
     position: absolute;
-    top: 10.5rem;
-    background-color: rgb(223, 223, 223);
+    top: 4rem;
+    background-color: var(--color-border);
     border-radius: 0 0 1rem 1rem;
+    z-index: 99;
     > div {
       padding-left: 1rem;
     }
     .DetailStationSerchDropdownSpace {
       padding-top: 1.3rem;
-      background-color: rgb(223, 223, 223);
+      background-color: var(--color-border);
     }
     .DetailStationSearchDropdownWapper {
-      width: 22.5rem;
-      max-height: 6rem;
+      width: 45vw;
+      max-height: 8rem;
       overflow: auto;
-      background-color: rgb(223, 223, 223);
-      border-radius: 1rem;
+      background-color: var(--color-border);
     }
     .DetailStationSearchDropdownButton {
       width: 100%;
-      background-color: rgb(223, 223, 223);
+      background-color: var(--color-border);
       border: none;
       border-top: 1px solid rgb(184, 184, 184);
       border-radius: 0 0 1rem 1rem;
@@ -56,7 +66,6 @@
       &:hover {
         background: lighten(rgb(223, 223, 223), 5%);
       }
-
       &:active {
         background: darken(rgb(223, 223, 223), 5%);
       }
@@ -64,8 +73,6 @@
   }
   .DetailStationLineFilter {
     width: 23rem;
-    padding-top: 8%;
-    padding-bottom: 4%;
     display: flex;
     justify-content: space-between;
     align-items: center;
